# apps/trio/openapi.yaml
openapi: 3.0.3
info:
  title: Merited trio — Commitment Signing, Token Mint + Conversion Verification, Net Settlement
  version: 0.1.0
  description: The security-critical trio (BUILD-SPEC §7). Phase 0 serves these contracts from behaviourally faithful simulators; PH1-24…26 replace them file-for-file behind the UNCHANGED TRIO-13 contract suite. This document is generated from the packages/contracts Zod schemas — CI fails on drift.
security:
  - serviceToken: []
components:
  securitySchemes:
    serviceToken:
      type: apiKey
      in: header
      name: X-Merited-Service-Token
      description: "Phase-0 transport gate (SYN-24): a shared service token on every route except /healthz. PH1-25 upgrades transport auth; payload trust always comes from signatures, never from this header."
...
paths:
  /trio/commitments:
    post:
      summary: Countersign a commitment (COR) — §7.1
      description: "Unsigned draft in, countersigned Commitment Object Record out. CORs are immutable: there is no update surface, and ledger roles revoke UPDATE/DELETE at the database. An optional budget registers a spend counter in Settlement; the COR itself stays budget-free (SYN-12). Emits `CommitmentCreated`."
      security: &a2
        - serviceToken: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                merchant_id:
                  type: string
                  pattern: ^mer_[0-9A-HJKMNP-TV-Z]{26}$
                offer_ref:
                  type: string
                  pattern: ^off_[0-9A-HJKMNP-TV-Z]{26}$
                bounty:
                  type: object
                  properties:
                    type:
                      type: string
                      enum:
                        - fixed
                        - pct_of_order
                    amount:
                      type: object
                      properties:
                        amount:
                          type: integer
                          minimum: 0
                        currency:
                          type: string
                          enum:
                            - GBP_pence
                      required:
                        - amount
                        - currency
                    pct_bps:
                      type: integer
                      minimum: 0
                      exclusiveMinimum: true
                  required:
                    - type
                take_rate_bps:
                  type: integer
                  minimum: 0
                agent_commission_bps:
                  type: integer
                  minimum: 0
                terms:
                  type: object
                  properties:
                    attribution_window_s:
                      type: integer
                      minimum: 0
                      exclusiveMinimum: true
                    eligible_identity_tiers:
                      type: array
                      items:
                        type: string
                        enum: &a1
                          - T1
                          - T2
                          - T3
                    max_conversions:
                      type: integer
                      nullable: true
                      minimum: 0
                      exclusiveMinimum: true
                    clawback_window_s:
                      type: integer
                      minimum: 0
                    valid_from:
                      type: string
                      format: date-time
                    valid_until:
                      type: string
                      format: date-time
                  required:
                    - attribution_window_s
                    - eligible_identity_tiers
                    - max_conversions
                    - clawback_window_s
                    - valid_from
                    - valid_until
                budget:
                  type: object
                  properties:
                    amount:
                      type: integer
                      minimum: 0
                    currency:
                      type: string
                      enum:
                        - GBP_pence
                  required:
                    - amount
                    - currency
              required:
                - merchant_id
                - offer_ref
                - bounty
                - take_rate_bps
                - agent_commission_bps
                - terms
      responses:
        "200":
          description: The countersigned COR.
          content:
            application/json:
              schema:
                type: object
                properties:
                  commitment:
                    type: object
                    properties:
                      commitment_id:
                        type: string
                        pattern: ^com_[0-9A-HJKMNP-TV-Z]{26}$
                      merchant_id:
                        type: string
                        pattern: ^mer_[0-9A-HJKMNP-TV-Z]{26}$
                      offer_ref:
                        type: string
                        pattern: ^off_[0-9A-HJKMNP-TV-Z]{26}$
                      bounty:
                        type: object
                        properties:
                          type:
                            type: string
                            enum:
                              - fixed
                              - pct_of_order
                          amount:
                            type: object
                            properties:
                              amount:
                                type: integer
                                minimum: 0
                              currency:
                                type: string
                                enum:
                                  - GBP_pence
                            required:
                              - amount
                              - currency
                          pct_bps:
                            type: integer
                            minimum: 0
                            exclusiveMinimum: true
                        required:
                          - type
                      take_rate_bps:
                        type: integer
                        minimum: 0
                      agent_commission_bps:
                        type: integer
                        minimum: 0
                      terms:
                        type: object
                        properties:
                          attribution_window_s:
                            type: integer
                            minimum: 0
                            exclusiveMinimum: true
                          eligible_identity_tiers:
                            type: array
                            items:
                              type: string
                              enum: *a1
                          max_conversions:
                            type: integer
                            nullable: true
                            minimum: 0
                            exclusiveMinimum: true
                          clawback_window_s:
                            type: integer
                            minimum: 0
                          valid_from:
                            type: string
                            format: date-time
                          valid_until:
                            type: string
                            format: date-time
                        required:
                          - attribution_window_s
                          - eligible_identity_tiers
                          - max_conversions
                          - clawback_window_s
                          - valid_from
                          - valid_until
                      merchant_sig:
                        type: string
                      platform_sig:
                        type: string
                    required:
                      - commitment_id
                      - merchant_id
                      - offer_ref
                      - bounty
                      - take_rate_bps
                      - agent_commission_bps
                      - terms
                      - merchant_sig
                      - platform_sig
                required:
                  - commitment
        "400":
          description: Malformed draft (`VALIDATION_FAILED`).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                    required:
                      - code
                required:
                  - error
        "401":
          description: Missing or wrong service token (`SERVICE_AUTH_FAILED`).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                    required:
                      - code
                required:
                  - error
  /trio/commitments/{id}/end:
    post:
      summary: End a commitment — §7.1
      description: 'Stops NEW token mints against the commitment. In-flight tokens still verify until their own windows close (SYN-34: §5.1 "old tokens still verify" governs — no retroactive repricing). Emits `CommitmentEnded`.'
      security: *a2
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Ended.
          content:
            application/json:
              schema:
                type: object
                properties:
                  commitment_id:
                    type: string
                    pattern: ^com_[0-9A-HJKMNP-TV-Z]{26}$
                  ended_at:
                    type: string
                    format: date-time
                required:
                  - commitment_id
                  - ended_at
        "404":
          description: Unknown commitment (`COMMITMENT_NOT_FOUND`).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                    required:
                      - code
                required:
                  - error
        "409":
          description: Already ended (`COMMITMENT_ALREADY_ENDED`).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                    required:
                      - code
                required:
                  - error
  /trio/commitments/{id}:
    get:
      summary: Commitment status — SYN-7
      description: Liveness (live / ended / expired / not_yet_valid) plus the Settlement counters (conversions used, budget remaining). Consumed by eligibility (CORE-6).
      security: *a2
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      responses:
        "200":
          description: Current status.
          content:
            application/json:
              schema:
                type: object
                properties:
                  commitment_id:
                    type: string
                    pattern: ^com_[0-9A-HJKMNP-TV-Z]{26}$
                  status:
                    type: string
                    enum:
                      - live
                      - ended
                      - expired
                      - not_yet_valid
                  conversions_used:
                    type: integer
                    minimum: 0
                  max_conversions:
                    type: integer
                    nullable: true
                    minimum: 0
                    exclusiveMinimum: true
                  budget_remaining:
                    type: object
                    nullable: true
                    properties:
                      amount:
                        type: integer
                        minimum: 0
                      currency:
                        type: string
                        enum:
                          - GBP_pence
                    required:
                      - amount
                      - currency
                required:
                  - commitment_id
                  - status
                  - conversions_used
                  - max_conversions
                  - budget_remaining
        "404":
          description: Unknown commitment (`COMMITMENT_NOT_FOUND`).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                    required:
                      - code
                required:
                  - error
  /trio/keys/merchant:
    post:
      summary: Issue a custodied merchant keypair — SYN-22
      description: "Custodied signing-key issuance (the §7 text omits this endpoint; the contract was added by MER-2 per SYN-22). Returns the key REFERENCE and public half only — private material never leaves trio custody (FakeSigner in Phase 0; KMS-enveloped Ed25519 in PH1-24). Idempotent: the reference is deterministic per merchant."
      security: *a2
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                merchant_id:
                  type: string
                  pattern: ^mer_[0-9A-HJKMNP-TV-Z]{26}$
              required:
                - merchant_id
      responses:
        "200":
          description: Key reference + public key.
          content:
            application/json:
              schema:
                type: object
                properties:
                  signing_key_ref:
                    type: string
                    minLength: 1
                  public_key:
                    type: string
                    minLength: 1
                required:
                  - signing_key_ref
                  - public_key
        "400":
          description: Malformed request (`VALIDATION_FAILED`).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                    required:
                      - code
                required:
                  - error
...